Refactor : 리뷰 반영 수정
