+++ src/main/java/com/zerobase/cafebom/product/service/dto/ProductDto.java
@@ -2,7 +2,7 @@ package com.zerobase.cafebom.product.service.dto;
 
 
 import com.zerobase.cafebom.product.controller.form.ProductForm;
-import com.zerobase.cafebom.product.domain.entity.SoldOutStatus;
+import com.zerobase.cafebom.product.domain.type.SoldOutStatus;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -12,7 +12,7 @@ public class ProductDto {
 
     private Integer id;
 
-    private Integer productCategory;
+    private Integer productCategoryId;
 
     private String name;
 
@@ -27,7 +27,7 @@ public class ProductDto {
     public static ProductDto from(ProductForm productForm){
         return ProductDto.builder()
                 .name(productForm.getName())
-                .productCategory(productForm.getProductCategoryId())
+                .productCategoryId(productForm.getProductCategoryId())
                 .description(productForm.getDescription())
                 .price(productForm.getPrice())
                 .soldOutStatus(productForm.getSoldOutStatus())
